backend/requirements.txt
fastapi         
uvicorn  
pydantic            
python-dotenv      
email-validator    
requests         
sqlalchemy
mysqlclient
resend
pymysql